+++ benchmark/src/BenchmarkSplayTree.cpp
@@ -29,7 +29,7 @@
 static void BM_SplayTree_Create_Average_Case(benchmark::State & state) {
 	std::random_device rd;
 	std::mt19937 gen(rd());
-	std::uniform_int_distribution<> dis(0, state.range(0));
+	std::uniform_int_distribution<> dis(0, static_cast<int>(state.range(0)));
 
 	forest::SplayTree <int, int> SplayTree;
 
@@ -49,7 +49,7 @@ BENCHMARK(BM_SplayTree_Create_Average_Case)->RangeMultiplier(2)->Range(1, 1 << 1
 static void BM_SplayTree_Search_Average_Case(benchmark::State & state) {
 	std::random_device rd;
 	std::mt19937 gen(rd());
-	std::uniform_int_distribution<> dis(0, state.range(0));
+	std::uniform_int_distribution<> dis(0, static_cast<int>(state.range(0)));
 
 	forest::SplayTree <int, int> SplayTree;
 
@@ -68,7 +68,7 @@ BENCHMARK(BM_SplayTree_Search_Average_Case)->RangeMultiplier(2)->Range(1, 1 << 1
 static void BM_SplayTree_Search_Worst_Case(benchmark::State & state) {
 	std::random_device rd;
 	std::mt19937 gen(rd());
-	std::uniform_int_distribution<> dis(0, state.range(0));
+	std::uniform_int_distribution<> dis(0, static_cast<int>(state.range(0)));
 
 	forest::SplayTree <int, int> SplayTree;
 
@@ -87,7 +87,7 @@ BENCHMARK(BM_SplayTree_Search_Worst_Case)->RangeMultiplier(2)->Range(1, 1 << 15)
 static void BM_SplayTree_Minimum_Average_Case(benchmark::State & state) {
 	std::random_device rd;
 	std::mt19937 gen(rd());
-	std::uniform_int_distribution<> dis(0, state.range(0));
+	std::uniform_int_distribution<> dis(0, static_cast<int>(state.range(0)));
 
 	forest::SplayTree <int, int> SplayTree;
 
@@ -106,7 +106,7 @@ BENCHMARK(BM_SplayTree_Minimum_Average_Case)->RangeMultiplier(2)->Range(1, 1 << 
 static void BM_SplayTree_Maximum_Average_Case(benchmark::State & state) {
 	std::random_device rd;
 	std::mt19937 gen(rd());
-	std::uniform_int_distribution<> dis(0, state.range(0));
+	std::uniform_int_distribution<> dis(0, static_cast<int>(state.range(0)));
 
 	forest::SplayTree <int, int> SplayTree;
